+++ src/core/scheduled/mod.rs
@@ -1,0 +1,5 @@
+mod xirr;
+mod xfv;
+
+pub use xirr::{xirr, xnpv};
+pub use xfv::xfv;
